// src/app/core/api/model/buildCacheInfoResponseDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BuildCacheInfoResponseDto { 
    pvcName: string;
    namespace: string;
    exists: boolean;
    phase: BuildCacheInfoResponseDto.PhaseEnum | null;
    /**
     * Allocated PVC capacity
     */
    capacity: string | null;
    storageClass: string | null;
    createdAt: string | null;
}
export namespace BuildCacheInfoResponseDto {
    export const PhaseEnum = {
        Bound: 'Bound',
        Pending: 'Pending',
        Lost: 'Lost'
    } as const;
    export type PhaseEnum = typeof PhaseEnum[keyof typeof PhaseEnum];
}